Restore the standalone entry point
The widget build swaps src/index.tsx for the web component entry and
restores it afterwards, keeping the original in src/index.bak.tsx while
it runs. The previous commit captured that swapped state: the standalone
entry contained the web component bootstrap, and the real one was
committed as the backup file.

A normal build would have compiled without an error and served an empty
page — the entry registered the custom element and never mounted the app
into #root. Nothing in lint, the unit tests or either build can catch
that, which is why it went out.

src/index.tsx is identical to develop again, the web component bootstrap
lives only in src/index-widget.tsx, and the backup file is gone.
